Changed initial connect calls for servers to be processed in batches of -maxsim size instead of waiting for each connect to complete.
This dramatically speeds up querying servers which use tcp connections, in testing
on 1000 servers reducing the time needed to query from 1min 15 seconds to under
4 seconds.

This new behaviour can be disabled by specifying -syncconnect

This also brings in a new state variable to each server, which can be used in the
future for more advanced processing instead of relying on server_name for state.

Added stripping of unprintable chars to xform_name, this can be enabled by specifying
-stripunprintable

Fixed trackmania name parsing.



git-svn-id: https://svn.code.sf.net/p/qstat/code/trunk@381 bac25679-d237-0410-bec6-f8029acc7ebe